requirements.txt
torch
torchvision
opencv-python
pillow
pandas
requests
customtkinter
